feat: Init frontend

+++ backend/app/schemas/cart.py
@@ -22,7 +22,7 @@ class CartItemUpdate(BaseModel):
 class CartItem(BaseModel):
     product_id: int
     name: str = Field(..., description="Product name")
-    price: int = Field(..., description="Product price")
+    price: float = Field(..., description="Product price")
     quantity: int = Field(..., description="Quantity in cart")
     subtotal: float = Field(
         ..., description="Total price for this item (price * quantity)"
